开启密码验证。 server支持暂停等待客户端连接 server执行线程逻辑优化

+++ common/src/main/java/com/fys/cmd/message/Cmd.java
@@ -28,13 +28,13 @@ public interface Cmd {
     void encoderTo(ByteBuf buf);
 
     //发送者的服务器端口
-    short getServerPort();
+    int getServerPort();
 
     //想要发送到的客户端端口
     String getLocalHost();
 
     //想要发送到的客户端端口
-    short getLocalPort();
+    int getLocalPort();
 
 
 }
