+++ common/src/main/java/com/fys/cmd/message/clientToServer/Pong.java
@@ -20,12 +20,12 @@ public class Pong implements Cmd {
     }
 
     @Override
-    public short getServerPort() {
+    public int getServerPort() {
         return 0;
     }
 
     @Override
-    public short getLocalPort() {
+    public int getLocalPort() {
         return 0;
     }
